+++ EFCore.PG.sln
@@ -31,8 +31,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EFCore.PG.Trigrams", "src\E
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EFCore.PG.FuzzyStringMatch", "src\EFCore.PG.FuzzyStringMatch\EFCore.PG.FuzzyStringMatch.csproj", "{86525F23-3EB0-4B0D-9BBB-55FF247D4DD6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EFCore.PG.Unaccent", "src\EFCore.PG.Unaccent\EFCore.PG.Unaccent.csproj", "{05C7487A-CF41-4A2E-9512-986687A04DFD}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -127,18 +125,6 @@ Global
 		{86525F23-3EB0-4B0D-9BBB-55FF247D4DD6}.Release|x64.Build.0 = Release|Any CPU
 		{86525F23-3EB0-4B0D-9BBB-55FF247D4DD6}.Release|x86.ActiveCfg = Release|Any CPU
 		{86525F23-3EB0-4B0D-9BBB-55FF247D4DD6}.Release|x86.Build.0 = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|x64.Build.0 = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Debug|x86.Build.0 = Debug|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|x64.ActiveCfg = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|x64.Build.0 = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|x86.ActiveCfg = Release|Any CPU
-		{05C7487A-CF41-4A2E-9512-986687A04DFD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -152,7 +138,6 @@ Global
 		{D7106D61-C7CA-4005-B31F-43281BB397AD} = {8537E50E-CF7F-49CB-B4EF-3E2A1B11F050}
 		{4DEDE46C-FABB-4AD3-A7DF-BF4A3AF00B06} = {8537E50E-CF7F-49CB-B4EF-3E2A1B11F050}
 		{86525F23-3EB0-4B0D-9BBB-55FF247D4DD6} = {8537E50E-CF7F-49CB-B4EF-3E2A1B11F050}
-		{05C7487A-CF41-4A2E-9512-986687A04DFD} = {8537E50E-CF7F-49CB-B4EF-3E2A1B11F050}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F4EAAE6D-758C-4184-9D8C-7113384B61A8}
