--- src/EFCore.PG.Unaccent/EFCore.PG.Unaccent.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <Authors>Kevin Martin</Authors>
-    <Description>unaccent module support plugin for PostgreSQL/Npgsql Entity Framework Core provider.</Description>
-    <PackageTags>npgsql;postgresql;postgres;Entity Framework Core;entity-framework-core;ef;efcore;orm;sql;unaccent;diacritics</PackageTags>
-
-    <AssemblyName>Npgsql.EntityFrameworkCore.PostgreSQL.Unaccent</AssemblyName>
-    <RootNamespace>Npgsql.EntityFrameworkCore.PostgreSQL.Unaccent</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\EFCore.PG\EFCore.PG.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-
-    <Compile Include="..\Shared\*.cs" />
-
-    <None Include="build\**\*">
-      <Pack>True</Pack>
-      <PackagePath>build</PackagePath>
-    </None>
-
-  </ItemGroup>
-
-</Project>
-
