+++ src/EFCore.PG/Query/ExpressionTranslators/Internal/NpgsqlMethodCallTranslatorProvider.cs
@@ -32,7 +32,8 @@ namespace Npgsql.EntityFrameworkCore.PostgreSQL.Query.ExpressionTranslators.Inte
                 new NpgsqlRangeTranslator(typeMappingSource, npgsqlSqlExpressionFactory),
                 new NpgsqlNetworkTranslator(typeMappingSource, npgsqlSqlExpressionFactory),
                 new NpgsqlJsonDomTranslator(typeMappingSource, npgsqlSqlExpressionFactory),
-                new NpgsqlJsonDbFunctionsTranslator(typeMappingSource, npgsqlSqlExpressionFactory)
+                new NpgsqlJsonDbFunctionsTranslator(typeMappingSource, npgsqlSqlExpressionFactory),
+                new NpgsqlUnaccentTranslator(npgsqlSqlExpressionFactory)
             });
         }
     }
